NoRef * Completed support for extract-year, extract-month etc. * Added support for sql DAY,WEEK,MONTH,YEAR functions in View entities.
According to http://blog.jooq.org/2015/02/25/how-to-extract-a-date-part-in-sql/, there isn't really a standard for the date functions. The following is true:


Extract(year from 
* Default 
* Cubrid
* Firebird
* H2
* HSQLDB
* Mariadb
* MySQL
* Postgres
* HANA
* Ingres
* Oracle (since jOOQ 3.6)

YEAR(
* Derby
* DB2
* Informix

DATEPART(
* ACCESS
* ASE
* SQLSERVER
* SYBASE

STRFTIME(
* SQLLITE


So using the above requires us to somehow determine which one is being required, before actually using it. Or you have to rely on actual functions for this altogether - which may complicate things during a select statement on large datasets.

git-svn-id: svn://data.ilscipio.com/ilscipio/ofbiz_foundation/trunk@9189 55bbc10b-e964-4c8f-a844-a62c6f7d3c80