OFBIZ patch: minilang: error handling: NO LONGER show success/event messages when there are errors
Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/453

The majority of services never want to show the event/success message
when there is an error, but minilang does not prevent this; meanwhile at
least 70 stock ofbiz services relied on call-service's default-message
to show a success message, but they are shown even when errors occur,
creating confusion.

This patch prevents success/event messages when errors occur (less
flexibility but more common sense).