Refs https://gitlab.ilscipio.com/SCIPIO-ERP/scipioce/issues/7 Fixed page permission error when recovering password.
When a password is recovered, a new one is send by email. If *password.email_password.require_password_change=true* in *security.properties*, the user couldn't login because the login process is not completed (therefore no userLogin in session) and the event returns "requirePasswordChange". This is correct but the *requirePasswordChange* screen def checks if *userHasAccount* is true, otherwise it throws the perm error. That part was wrong, so now, when the login process goes to "requirePasswordChange" a *tmpUserLogin* is set as a request attribute, so the screen can check tmpUserLogin.requirePasswordChange="Y" then render the screen properly even if *userHasAccount* is false.