all: replace javolution FastSet, FastMap, FastList + extra
Replaces the javolution.util.* collection types in all files with the
following native Java types (directly or through factory/util methods):
* FastMap -> HashMap UNLESS iter order needed, then LinkedHashMap
* FastList -> ArrayList
* FastSet -> HashSet

Specific factory methods, expressions and utilities used:
* *.java: UtilMisc#newMap, #newInsertOrderMap, #newList, #newSet
* *.groovy: [:], [], [] as Set, new ArrayList(c), new HashMap(c).
* others (minilang): as appropriate

This commit is a FIRST part mass-conversion and covers all code, but
does not include addons or EntitySaxReader. A SECOND part might or might
not occur to standardize the UtilMisc calls after they are further
reviewed for type choice (low priority).

This work is similar to but heavily modified from: OFBIZ-5781,
OFBIZ-5169, OFBIZ-6298. Notably, the default List type in Scipio chosen
is ArrayList, not LinkedList. ArrayList is faster and consumes less
memory in the majority of cases when considering the usages in Scipio
(and Ofbiz). It is also the List type in Groovy ([]). Compared to
FastList from javolution 5.x, ArrayList is now often faster, while
LinkedList is slower.

This commit also contains EXTRA fixes:
* concurrency bugfix for static cache in ValueLinkApi
* concurrency bugfix for static cache in ContentManagementWorker
* removed needless memory hog vars in
ServiceArtifactInfo#writeServiceCallGraphEoModel
* ArrayList initial capacity for WebToolsServices#getEntityRefData,
#exportEntityEoModelBundle

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/480