security: OFBIZ merge: 1759065 from 17.12 (content services allow-html)
1 commit: "In rev. 1536324  Jacques Le Roux changed the value of the
attribute allow-html from "safe" to "any" and deprecated the usage of
the "safe" value. However the "safe" behavior is necessary to allow
these fields to accept tags in a safe way: for this reason that commit
should be reverted and a different solution should be implemented. In
the meantime, this commit changes the allow-html values for the blog
entries to "none". [CVE-2016-6800]"

NOTE: for Scipio, this is very low-urgency fix since these services are
not in use

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/538