OFBIZ patch: logging: Debug.java: omit the "Groovy" from the output module name
This was of my making, but it simply takes up space for nothing and
looks poor, while the script names are usually clear even without
reference to groovy. So the new workaround is to simply omit the
extension (not possible to include it due to the way log4j
works/isused).