start: loaders: new scipio-events.properties with startup finish hooks for all components
Adds support for a new scipio-events.properties file that can be
included by any component to define an event handler for startup finish
callback. In addition, scipio-events.properties can now be used instead
of the 6 different ofbiz loader files as a simple way to add startup
hooks without having to modify 7 files or deal with consecutive
numbering.

This will make it easier for anyone including stock Scipio, addons, and
hot-deploy components to add code to run on startup without being forced
to define jobs in the DB, and avoiding any build dependency issues.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/524