all: replace UtilMisc.newList with LinkedList<...>
This in effect REVERTS the UtilMisc.newList ArrayList usage mainly in
applications/ classes, in order to minimize differences with upstream.
These instances of LinkedList are not in performance-critical areas so
they are negligible gain; so instead the ArrayList will be added to
performance-critical areas in Scipio of these peripheral classes;
meanwhile these were heavily negatively influencing maintenance through
massive diff noise.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/480