security: ofbiz merge: aa8dbc3207 from 17.12 (X-XSS-Protection - HTTP headers)
1 commit: "X-XSS-Protection This header enables the Cross-site scripting
(XSS) filter built into most recent web browsers. It's usually enabled
by default anyway, so the role of this header is to re-enable the filter
for this particular website if it was disabled by the user. This header
is supported in IE 8+, and in Chrome (not sure which versions). The
anti-XSS filter was added in Chrome 4. Its unknown if that version
honored this header. FireFox has still an open bug entry and "offers"
only the noscript plugin.
https://wiki.mozilla.org/Security/Features/XSS_Filter
https://bugzilla.mozilla.org/show_bug.cgi?id=528661"

from https://github.com/apache/ofbiz-framework.git branch release17.12

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/471