ofbiz patch: all: makeLink secure+encode flag fixes in *.java
Switched many instances that passed hard secure=false to secure=null;
and in some of these, also switched from encode=false to encode=true.

The secure=null is for prevention and future changes, because
secure=false was already ignored in many of these cases.

However, encode=false was a problem,
because using encode=false is a bad idea in most of these calls because
it is now less associated with jsessionid (almost always omitted) and
more with URL rewriting.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/31