ofbiz patch: SOAPClientEngine: new context handler, enabled sharing
Enables sharing of the axis2 ConfigurationContext for the SOAP engine
and the Scipio Camel addon built-in connector; dramatically increases
SOAP call time and removes verbosity.

Handled using a new SOAPContextHandler in part to delay the init and to
prevent high-level callers of invokeRemoteMirrorService (like camel)
from having to handle low-level axis2 objects.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/460,
https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/463