ofbiz patch: EntityUtilProperties: new SystemProperty.useEmpty flag + restore legacy empty handling
This patch:
* adds a new SystemProperty.useEmpty flag that when set to Y causes
empty value to become significant (ofbiz 16+ default behavior)
* restores the ofbiz 14/15 default SystemProperty.systemPropertyValue
empty value handling; considered NON-significant by default (meaning
SystemProperty.useEmpty default is N) [this behavior was temporarily
inverted in last commit bca66c915990dbdb27bf756b04ec382fb6710d43]

The reason for preserving the old behavior as default is that the
SystemProperty records contain useful information and it is an
anti-pattern to delete them in order to fallback on *.properties files.

However, it must still be possible to set empty values.

In Scipio, the useEmpty flag is not yet managed by any code, but it will
be up to the updating/UI code in future to manage it.

This change ensures preservation of backward-compatibility.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/613