ofbiz patch: NotificationServices: setBaseUrl: add baseWebappPath in context
New context field "baseWebappPath" for email templates, which contains
webappPathPrefix+contextRoot for the current website (only).

NOTE: this is mainly a workaround; future changes will hopefully
eliminate this...

DEV NOTE: it has been found impossible to include webappPathPrefix in
baseUrl/baseSecureUrl (this was the initial workaround, but had to be
reverted).

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/607