UrlFilterHelper: use target webapp info instead of current request, new attributes
The urlrewrite.xml files should not rely on HttpServletRequest to
determine the current webapp anymore, because it could differ depending
on the implementation of ScipioUrlRewriter used for inter-webapp links.

Instead some new attributes are populated from the target webapp
determine from context.

Also, the new attributes are higher-level, so features may be added to
UrlFilterHelper more easily.