ofbiz patch: build: add component arg to all load-demo invocations
Allows specifying single component arg using command line on the load
targets, e.g.:

  ./ant load-demo -Dcomponent=my-component

This was not exposed in stock in ant build; it was previously only
possible to run by invoking the ofbiz.jar directory. But we need to
avoid direct invocation of the jar in Scipio (user-unfriendly and
bypasses too much).