ofbiz merge: c692cfeb61 from trunk (GenericValue tests)
1 minor commit (c692cfeb617fb9d41289190573681f13d94017d9):
"Generic value is null or not should be checked using native checks.
Utility method isEmpty or isNotEmpty is not needed to check if generic
value is null or not  because generic values can never be empty. They
can be null or not nulls. So made changes accordingly all over the
places. (OFBIZ-8471)"
from https://github.com/apache/ofbiz-framework.git trunk
-e 
Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/626