ofbiz merge: partial 577423c3145 from trunk (JobManagerLock feature, disabled)
1 minor commit, all UI changes commented out:
"Add the possibility to set a blackout period for job manager service.
During this time all async and scheduled service are put on hold.
This improvement add a entity JobManagerLock, a new screen to add/edit
job manager lock and an information for end user on FindJob and
ScheduleJob screen.
See more information on issue OFBIZ-6790"
SCIPIO: I committed the code and entities but commented it out of the UI
- the change is here for future use and synchronizing the codebase only
for now.

Ofbiz-Commit: 577423c3145800eda6926e75c531ceb960e04014
Ofbiz-Repo: https://github.com/apache/ofbiz-framework.git
Ofbiz-Branch: trunk

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/632