ofbiz merge: c5f06e06869 from trunk (statusEndDate EntityAutoEngine+FinAccountStatus patch)
1 commit:
"Improve FinAccountStatus entity to manage changeByUserLoginId field
with conversion of minilang service to entity-auto. This previous
service manage a specific case for the statusEndDate field that set the
date when a new entry is stored on statusDate and the same value on
previous stored on statusEndDate. This field is also present on
ExampleStatus, so extend the EntityAuto engine to manage this case.
OFBIZ-7623"

Ofbiz-Commit: c5f06e06869bd7af463fb0754247524b89f7e0df
Ofbiz-Repo: https://github.com/apache/ofbiz-framework.git
Ofbiz-Branch: trunk

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/632