ofbiz merge: 420ccc94040 from trunk (remove all call-bsh instances in project, contains bugfixes)
1 commit:
"OFBIZ-7765 converted Minilang code that was using the deprecated
"call-bsh" element to use the "script" element with Groovy. Thanks to
Deepak Dixit for the work and the patch and to Nicolas Malin for the
review and imporvement."
SCIPIO: IMPORTANT: beanshell had issues with varargs methods as now
heavily used in UtilMisc and elsewhere, so this change is also a bugfix
for any bsh code that invokes a varargs method.

NOTE: this does not remove the call-bsh element from the *.xsd; keeping
backward compatibility for now or until it becomes a serious issue (e.g.
security issue).

Ofbiz-Commit: 420ccc94040a1875e424f2d690f4d5fb1f6eb178
Ofbiz-Repo: https://github.com/apache/ofbiz-framework.git
Ofbiz-Branch: trunk

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/556,
https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/641