stock patch: widget: new (Widget)RenderOptions, warn missing renderers
Added new (Widget)RenderOptions class, because doing several MapStack
lookups for frequent ops like logging statements is much too heavy, and
this was sorely missing.

Also adds initial support for specific rendering contexts to warn when
string renderers are missing; by default this is only logged in verbose,
which is correct because optional depending on context, but we have one
or more circumstances where is important to warn because should not
happen or to inform clients.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/645