security: entity: crypto: REMOVE -kek-old (insecure DES key gen)
Removes -kek-old command line arg from org.ofbiz.base.crypto.Main. Fixes
one or two warnings.

DES is insecure and the remaining DES code in Scipio, should not be used
to encrypt new records only to decrypt existing ones.

Since ValueLink is not working anymore anyway, there is no reason to
leave this in.

For emergency testing (ONLY), the code is left commented in Main.java
and could be uncommented.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/489,
https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/488,
https://github.com/ilscipio/scipio-erp/issues/6