stock patch: ScriptUtil: add bsh scriptlet backward-compat mode (groovy), warnings
This adds a best-effort beanshell compatibility mode for old expressions
and scriptlets that may still use the "bsh:" prefix or language name.

The code tries to run Groovy in BSH-like variant emulation mode (with no
exceptions on missing properties).

Also adds log warnings to incite all code to switch to groovy as soon as
possible.

This is part of Beanshell (bsh) deprecation in favor of Groovy.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/641