stock patch: entity: filterByDate: do not throw ex. on bad field names
Patches the most common date filter utils, EntityQuery.filterByDate and
EntityUtil.filterByDate, so that they log a detailed error instead of
throwing it, and return the data without filtering.

This is because due to evolving schema, accidentally adding a
filterByDate call on entity without fromDate/thruDate is a very common
error, and in 90% of the cases the solution is simply removing the call,
which has the same effect as this change.

This change prevents causing fatal errors in such cases, which are
damaging for no reason.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/682