ofbizContentUrl/ContentUrlTag: appendContentPrefix: try more fallbacks
When rendering from static context (e.g. emails), if no explicit content
prefix is configured in url.properties or WebSite, now falls back on
baseUrl/baseSecureUrl or (if these missing in context) the host
configured for nav links in url.properties or WebSite.

Caveat: Unfortunately as a side effect, the no-prefix-configured warning/error is never shown anymore, because localhost is always used as last-case fallback; but this is easily seen.