services.xsd: new attribute: type-convert="[true|false]"
As a new feature in Scipio, an IN attribute can now be tagged
type-convert="true" so that type conversion will always be applied for
that attribute when LocalDispatcher.runSync/runAsync is called, even
outside of service event handler.

This can be used in some cases to fix compatibility issues and trivially
generalize a service.

TODO: For now this only supports surgical conversions on individual
attributes in order to avoid abuse, but in future per-service defaults
and overrides could be supported.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/830