stock patch: createWorkEffortGoodStandard: estimatedQuantity type-convert="true"
(New solution)

For createWorkEffortGoodStandard and updateWorkEffortGoodStandard
services, several stock code callers were passing BigDecimal for
estimatedQuantity though it is a Double in the interface; this was an
error (which was hidden by service engine bug #830).

This all-compatible change retroactively makes those errors to be
non-errors by automatically type-converting the estimatedQuantity field.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/830