stock patch: UtilProperties: getMessage now support optional (return "")
This has been missing for a very long time; returning the key name only
ever made sense for the UI, not for internal code; now pass
optional=true to return "" when no message AND prevent the logging.

The stock logic of comparing the key name to the value to test if change
is inherently flawed (the key name return was designed for straight UI
usage) and this will slowly help to phase it out over time.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/833