stock patch: ServiceDispatcher: add compat mode for old perm'd services
Adds Compatibility (no-crash) mode for old services having permission
services.

Due to a very old upstream bug, old services which defined a
permission-service in their interface (only)
automatically received makeValid calls on their contexts, which they
were not supposed to;
the consequence is that many services were developed and tested without
being properly type- and context-validated.
When set to true, the service dispatcher will preserve the old behavior
of apply an automatic makeValid on the contexts
passed to such services, but it will log violations as errors in
error.log so these instances can be found without
breaking any processes.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/830