order email notifications: new store email default WebSite fallback
Following new configuration in catalog.properties and
ProductStoreWorker, OrderHeaders missing an explicit webSiteId  now send
emails to the default configured WebSite (if configured).

There are several reasons for this:
* orderentry previously did not populate OrderHeader.webSiteId at all,
so this is a form of backward-compatibility
* For orders not placed on a website (such as by phone), it is debatable
whether OrderHeader.webSiteId itself should be populated; but usually it
is regardless still desirable to send email notifications if the buyer
provided an email address.

NOTE: This change means that simply having OrderHeader.webSiteId empty
is not enough to prevent emails anymore; the OrderHeader.webSiteId field
was always more generic than an "email control" field and should not be
relied on to prevent emails.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/684