stock patch: sendMailFromScreenInterface: clarify webSiteId
Now reads:
"The webSiteId of the WebSite this email should appear to be from, or
in other words - from the point of view of the screen - the "current"
WebSite.
Used for building links to the web site, themeing, etc.
NOTE: This is REQUIRED for store frontend emails (WARNING: Cannot be
enforced! Caller must ensure).
Backend emails can leave this empty."

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/684