appl: *.ftl: refactor common includes to promote new *lib.ftl #imports
Due to heaviness of and problems with the #include directives in
applications, this creates new *lib.ftl files in applications webapp
common folders in which some utilities have been moved, and which can be
imported (the proper Freemarker way) using #import, such as:
  <#import "component://order/webapp/ordermgr/common/orderlib.ftl" as
orderlib>

To keep naming consistent and files easy to remember and see in IDE, the
namespace name is encouraged to be the lib filename. It is recommended
to keep the lib names short (11 chars total or lower) since freemarker
IDE is limited.

This changes fixes the duplicate initialization problems (without having
to manually code include guards, which are slower) and overhead and
makes things easier to locate.

In addition:
* Implicitly fixes duplicate includes (due to #import)
* Removes a number of useless *common.ftl files in applications
* Standardizes the include/import to always use "component://" so
copy-pasting results in fewer errors

Applications that have not been refactored (much):
* shop - Due to compatibility and heavy presence of markup in the common
macros, as well as client code simplicity (only minor acctlib include
changes)

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/849