sendMailFromScreen: new webSiteId+productStoreId handler/impl override
This inserts a new sendMailFromScreen service override in commonext
component, which with autoInferParams (default: true), now attempts to
automatically determine webSiteId and productStoreId using each other
and using orderId.

This is a rather intrusive central patch to solve the problem of many
email services forgetting to pass these necessary parameters; it is
simply too unreliably to patch them all one-by-one.

This behavior can be disabled by passing autoInferParams=false, but
it is unlikely to cause any problems for most emails.

Additional changes:
* EmailServices.sendMailFromScreen recognizes webSiteId on its own too.
* Various fives

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/684