security: screen renderer: no TRANSACTION_TIMEOUT from request params
This changes the framework default behavior of accepting a
TRANSACTION_TIMEOUT from request parameters; it must now be requested
explicitly in screen definitions. The TRANSACTION_TIMEOUT is still
read from request/session/servletcontext attributes for the time being,
though (WARNING) this behavior is also being investigated.

NOTE: The added elements in widget-screen.xsd are for testing purposes
or to provide a quick-fix for any unexpected compatibility breaks
from this change (unlikely); they are labeled EXPERIMENTAL
and may be subject to change if any further issues.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/909