stock patch: Delegator+FTL: new exception handling and methods for *.ftl
Adds the following methods to Delegator interface for implementation
purposes (a few more than are ideal, but it's the only way to make this
work without losing access to functionality):
* fromSafe/fromUnsafe
* selectSafe/selectUnsafe
* querySafe/queryUnsafe

NOTE: These are NOT intended for ordinary use - rather the simpler
query(), from(...) and select(...) should be almost always used for that
- but rather they're used to implement DelegatorWrapperModel, which for
FreeMarker templates switches the behavior to prevent
GenericEntityException throws.

In addition, this sets the FTL env. "EntityQuery" var to the
EntityQuerySafe class.

NOTE: This is technically a compatibility-affecting change, however 99%
of known FTL code is not affected, and the previous behavior was not
very acceptable.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/915