stock patch: shop: CustomerSurvey.groovy: remove broken callResult flag
This flag prevents purchasing the same gift card twice with an account
(e.g. DemoCustomer) because the existing response is taken by the flag
as a result to prevent filling in a new response, i.e. it causes
Survey.allowMultiple Y to be ignored.

NOTE: It is unclear if any original cases needed this flag, but
currently for Scipio it is more damaging than good. May be revisited
later.

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/844