orderentry: showcartitems.ftl: replaced survey link with modal
the stock survey link is beyond confusing: ViewSurveyResponses shows all
responses for all orders for a survey, while EditSurveyResponse actually
creates a new response - it doesn't permit editing the response for the
order.

this provides a simple modal instead, for the time being, using a
wrapper around genericresult.ftl (may improve in future).

Refs https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/958