MapContext: scpCtxAutoClose: new auto object closing callback (context field)
Collection of new utils that will permit to add easy patches
for many of the EntityListIterator problems. the most generic workaround
is a new context field scpCtxAutoClose whose objects are closed
on pop() - it is built into MapContext so it simply works everywhere,
and it doesn't require use of MapContext interface, but from java
can use new AutoClose and MapContext helpers to save from boilerplate.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/1052
https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/1054