stock patch: GenericEntity: new member: dataOrigin (get/setDataOrigin)
Adds an extra entity field used to record the logical source
of the data in the entity. It is separate and distinct from the
entity engine (delegator) data source, so as to not require any java
implementation to use or set. Eliminates the need to pass around
extra variables to record the origin of the data in an entity.
to be used for solr/solrProduct.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/issues/1102