ShoppingCartFactory: fixed createShoppingCartHelper to createWebShoppingCartHelper, simpler return values
Classically ShoppingCartHelper is only used in the web context, but it can be used elsewhere, and
the distinction has to match the other overloads otherwise client code that needs a different ShoppingCartHelper
will eventually break when used in different case.

This also generifies the return values so configured client code doesn't need to cast all the time.