entity: TransactionUtil: suspend(): do not warn when no transaction (verbose only by default)
The warning in this method burdens all client code with prevent the warning which is
very rarely needed so only complicates client code. Now you can basically put the suspend/resume
pattern everywhere so it always works without worrying, so the resulting code is more
reusable and less error prone and simpler for clients.

It's still possible to call suspendOrWarn if needed and use verbose mode to debug.