EntityListCache: stored entity list is now unmodifiable (WARNING: callers may need fixes!)
The entity list cache failed to protect stored lists and callers may have been inadvertently
editing cache contents unaware that this usage was an error.

This change reflects the existing behavior of making the GenericValues unmodifiable.
It is not a good idea to return list copies because they can be huge, 99% of everything does
not need copies so it all slows down for a few cases, and the entities themselves are already
enforced read-only. Buggy callers must be fixed.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1177