productImageAutoRescaleAll: added dedicated service + abortProductImageAutoRescaleAll, allResumeId
productImageAutoRescaleAll now can be called with mostly sane defaults.
abortProductImageAutoRescaleAll can be called to stop productImageAutoRescaleAll as
it may take a long time to run.

allResumeId may be set to the last seen productId so that you can manually resume
from another product. For this, the default is to order by productId.

NOTE: createSizeTypeContent is default false - must pass true if expect creation of
new data ProductContent/DataResource records for missing siteTypes.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1233