contentImageFileScaleInAllSizeCore: stock workaround for imageOrigUrl not matching configured prefix in catalog.properties
Some clients have data that is hardcoded or does not match image.url.prefix configuration,
so this attempts to preserve per-image pre-existing extra prefix before the coded prefix, if possible.
Unfortunately ofbiz design does not record the on-disk location of images for product images.
WARN: Only works if the image.url.prefix is contained within the imageOrigUrl.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1233