DatabaseUtil: check-indices-on-start: added postgres support for index field changes (check-modified-indices-on-start)
Postgresql-only: when check-indices-on-start="true" either in entityengine.xml
or on command line with -Dscipio.entity.datasource.localpostnew.check-indices-on-start=true
modified indices (except auto indices on TX fields) now attempt to detect if any fields
changed; the index is deleted and recreated with the new fields and the changes
are logged.

Due to limited SQL support this currently only works in postgres.
NOTE: Deleting an index definition does not remove it from the database.

If problems occur with the feature and you don't want to disable check-indices-on-start
completely, the new functionality can be disabled per-datasource using:
    check-modified-indices-on-start="false"

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1204