ProductImageVariants: support sizes missing ProductContentType entries
The size's ProductContentType can often be temporarily missing as ImageProperties.xml
are updated or the seed data is not created and an image is rendered in the frontend.
This allows the code to work temporarily until they're created by
productImageAutoRescaleAll or other image generation service.

If the size's ProductContentType is missing (which may happen if there are no image
instances of it in data), it is now generated in-memory to avoid negatively
affecting rendering.

Also applies misc. fixes.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1290