view-entity: added support for alias-columns="true/false" and select="false" on aliases
entityengine.xml defines alias-view-columns for each db but is flatlined to false in
all cases; alias-columns allows to enable the feature per-view-entity as needed without
causing errors globally.

In order for this feature to work, SQL rules must be respected which means columns
must be excluded that are not group-by because this is not implied, so a select="false"
can be added to <alias/> elements to prevent this.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1300
https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1301