UtilCache: defer cache line execution pulse subscription to after adding to memoryTable
An error existed in upstream code where the cache lines were added as execution pulses
for delayed removal before adding to memoryTable; this made it technically possible for
removes to occur extremely quickly before the cache line was added causing tiny memory
leaks. In addition, this change in putIfAbsentInternal means that cache lines that
are not added to memoryTable don't need to be added as execution pulses at all, which
may save a small amount of overhead.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1318