generateProductAlternativeUrlsCore: don't sets textData to empty
For data safety reasons, these services no longer allow setting
the textData of main and localized ALTERNATIVE_URL content records
to empty, even if an invalid old value is retained instead.

In principle in these cases records might be deleted, but this
is safer for the time being, for future review.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1322