entity-eca: added support for reload-value for store/return operation
store/return entity ECA service calls place a GenericValue as value-attr
which may contain only the fields to be updated rather than the entity
value from the data source. Setting reload-value="true" instructs the
ECA handler to reload the value from data source before passing it
as value-attr to the service, so that it's not necessary to edit
services to achieve this.

This may even be used for async services even though the value
may not be visible at data source level in the new transaction.

https://gitlab.ilscipio.com/scipio-dev/dev/scipioce-dev/-/issues/1322